♻️ refactor: fetch, type 파일 분리
