+++ src/app/MainProductLists.tsx
@@ -1,20 +1,16 @@
 'use client';
 
 import ProductCard from '@/components/ProductCard';
-import {
-  getLikeProducts,
-  getRecipes,
-  addBookmark,
-  deleteBookmark,
-  getLikeRecipe,
-} from '@/data/functions/post';
 import { ApiRes, LikeItemType, ProductType } from '@/types';
-import { Post } from '@/types/post';
+import { Post } from '@/types/recipe';
 import useUserStore from '@/zustand/useStore';
 import useBookmarkStore from '@/zustand/useBookmarkStore';
 import Link from 'next/link';
 import { useEffect, useState } from 'react';
 import RecipeCard from './recipe/RecipeCard';
+import { getLikeProducts } from '@/data/functions/product';
+import { getLikeRecipe, getRecipes } from '@/data/functions/recipe';
+import { addRecipeBookmark, deleteRecipeBookmark } from '@/data/actions/recipe';
 
 interface MainProductListsProps {
   products: ProductType[];
@@ -105,10 +101,10 @@ export default function MainProductLists({ products }: MainProductListsProps) {
 
     try {
       if (isBookmarked) {
-        const res = await deleteBookmark(accessToken, bookmarkId);
+        const res = await deleteRecipeBookmark(accessToken, bookmarkId);
         if (res.ok === 1) remove(postId);
       } else {
-        const res = await addBookmark(accessToken, postId);
+        const res = await addRecipeBookmark(accessToken, postId);
         if (res.ok === 1 && res.item) add(postId, res.item._id);
       }
     } catch (err) {
