+++ src/app/page.tsx
@@ -1,9 +1,9 @@
 import './main.css';
 import MainSlide from '@/app/MainSlide';
 import ValueSlide from '@/app/ValueSlide';
-import { getProducts } from '@/data/functions/post';
 import { ProductType } from '@/types';
 import MainProductLists from '@/app/MainProductLists';
+import { getProducts } from '@/data/functions/product';
 
 export default async function Home() {
   const res = await getProducts();
