+++ src/app/recipe/RecipeCard.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import { Bookmark } from 'lucide-react';
-import { Post } from '@/types/post';
+import { Post } from '@/types/recipe';
 import useBookmarkStore from '@/zustand/useBookmarkStore';
 
 interface RecipeCardProps {
