+++ src/data/actions/product.ts
@@ -1,7 +1,7 @@
 'use server';
 
 import { ApiRes, ApiResPromise, LikeItemType } from '@/types';
-import { revalidateTag } from 'next/cache';
+import { revalidatePath, revalidateTag } from 'next/cache';
 
 const API_URL = process.env.NEXT_PUBLIC_API_URL;
 const CLIENT_ID = process.env.NEXT_PUBLIC_CLIENT_ID || '';
@@ -86,6 +86,54 @@ export async function productDeleteLike(
     revalidateTag(`bookmarks`);
     revalidateTag(`/shopping/${_id}`);
     revalidateTag(`/shopping`);
+  }
+
+  return data;
+}
+
+/**
+ * 상품 좋아요 삭제 (마이페이지용)
+ * @param {ApiRes<PostReply> | null} state - 이전 상태(사용하지 않음)
+ * @param {FormData} formData - 삭제할 댓글 정보를 담은 FormData 객체
+ * @returns {Promise<ApiRes<PostReply>>} - 삭제 결과 응답 객체
+ * @description
+ * 댓글을 삭제하고, 성공 시 해당 게시글의 댓글 목록을 갱신합니다.
+ */
+export async function deleteLike(
+  state: ApiRes<LikeItemType> | null,
+  formData: FormData,
+): ApiResPromise<LikeItemType> {
+  const _id = formData.get('_id');
+  const accessToken = formData.get('accessToken');
+
+  let res: Response;
+  let data: ApiRes<LikeItemType>;
+
+  const body = {
+    target_id: 'any',
+  };
+
+  try {
+    res = await fetch(`${API_URL}/bookmarks/${_id}`, {
+      method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+        'Client-Id': CLIENT_ID,
+        Authorization: `Bearer ${accessToken}`,
+      },
+      body: JSON.stringify(body),
+    });
+
+    data = await res.json();
+  } catch (error) {
+    // 네트워크 오류 처리
+    console.error(error);
+    return { ok: 0, message: '일시적인 네트워크 문제가 발생했습니다.' };
+  }
+
+  if (data.ok) {
+    revalidateTag(`bookmarks`);
+    revalidatePath(`/mypage/likes`);
   }
 
   return data;
