+++ src/data/functions/user.ts
@@ -1,0 +1,22 @@
+import { ApiResPromise, MemberType } from '@/types';
+
+const API_URL = process.env.NEXT_PUBLIC_API_URL;
+const CLIENT_ID = process.env.NEXT_PUBLIC_CLIENT_ID || '';
+
+// 단일 회원 정보 조회
+export async function getMember(_id: number): ApiResPromise<MemberType> {
+  try {
+    const res = await fetch(`${API_URL}/users/${_id}`, {
+      headers: {
+        'Client-Id': CLIENT_ID,
+      },
+      next: {
+        tags: [`users/${_id}`],
+      },
+    });
+    return res.json();
+  } catch (error) {
+    console.error(error);
+    return { ok: 0, message: '일시적인 네트워크 문제로 조회에 실패했습니다.' };
+  }
+}
