+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './user';
 export * from './product';
 export * from './api';
+export * from './recipe';
